Building OSX for arm is now correct
- bash 3.x support for scripts.
- Info.plist now has generated version numbers.
- Info.plist now has correct icons.
- libgit2 now builds correctly for OSX M1. It is not included correctly in libgit2sharp.

Bonus: now defaulting to current OS when building or publishing.

+++ run.sh
@@ -10,7 +10,6 @@ usage() {
 }
 
 # Change to expected directory. Everything is relative from here.
-ORIGINAL_DIR="$(pwd)"
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 ROOT_DIR="$SCRIPT_DIR"
 cd "$ROOT_DIR" || exit 1
@@ -20,11 +19,11 @@ build_usage() {
   echo ""
   echo "$0 build [OPTIONS...]:"
   echo "OPTIONS:"
-  echo " -a, --arch <arch>      Sets the architecture when publishing Is one of x86, x64, arm, arm64. (default = x64)"
+  echo " -a, --arch <arch>      Sets the architecture when publishing Is one of x86, x64, arm, arm64. (default = current)"
   echo " -c, --config <config>  Sets the configuration. Is one of debug, release, or lint. (default = debug)"
   echo " -q, --quiet            Verbosity is set to quiet."
   echo " --publish              Publishes build."
-  echo " --os OS                Sets the os to compile to. linux or win. (default = win)"
+  echo " --os OS                Sets the os to compile to. linux or win. (default = current)"
   echo " --single-file          Publishes a single file."
 }
 
@@ -34,14 +33,25 @@ build_cmd() {
     PUBLISH="0"
     OS="win"
     ARCH="x64"
-    VERSION=""
+    VERSION="0.0.0.0-internal-build"
+    VERSION_SHORT="0.0.0.0"
     
     SHARED_COMPILATION=""
+
+    case "$(uname -s)" in
+        Darwin)     OS="osx";;
+        MINGW64*)   OS="win";;
+    esac
+
+    case "$(uname -m)" in
+        arm64)  ARCH="arm64";;
+        x86_64) ARCH="x64";;
+    esac
 
     while [[ $1 != "" ]]; do
         case "$1" in
             --arch|-a)      shift
-                            case "${1,,}" in
+                            case "$(echo "$1" | tr '[:upper:]' '[:lower:]')" in
                                 arm)            ARCH="arm";;
                                 arm64)          ARCH="arm64";;
                                 x86)            ARCH="x86";;
@@ -53,7 +63,7 @@ build_cmd() {
                             esac
                             ;;
             --config|-c)    shift
-                            case "${1,,}" in
+                            case "$(echo "$1" | tr '[:upper:]' '[:lower:]')" in
                                 debug)          CONFIG="Debug";;
                                 release)        CONFIG="Release";;
                                 lint)           CONFIG="Lint";;
@@ -68,7 +78,7 @@ build_cmd() {
             --publish)      PUBLISH="1"
                             ;;
             --os)           shift
-                            case "${1,,}" in
+                            case "$(echo "$1" | tr '[:upper:]' '[:lower:]')" in
                                 win|windows)    OS="win";;
                                 linux)          OS="linux";;
                                 osx|mac)        OS="osx";;
@@ -80,6 +90,7 @@ build_cmd() {
                             ;;
             --version)      shift
                             VERSION="$1"
+                            VERSION_SHORT=${VERSION/-*}
                             ;;
             --ci)           SHARED_COMPILATION="-p:UseSharedCompilation=false"
                             ;;
@@ -90,6 +101,8 @@ build_cmd() {
         shift
     done
     
+
+    BUILD_DIR="$ROOT_DIR/bin/$CONFIG/net6.0"
     if [[ "$PUBLISH" == "1" ]]; then
         PUBLISH_ARGS=("--self-contained=false" "-p:PublishSingleFile=true")
         
@@ -101,23 +114,33 @@ build_cmd() {
         if [[ "$VERSION" != "" ]]; then
             PUBLISH_ARGS+=("-p:Version=$VERSION")
         fi
-        
+
         dotnet publish -c "$CONFIG" "${VERBOSITY[@]}" "$SHARED_COMPILATION" --arch "$ARCH" --os "$OS" "${PUBLISH_ARGS[@]}"
-    
-        PUBLISH_DIR="$ROOT_DIR/bin/$CONFIG/net6.0/$OS-$ARCH"
+
+        BUILD_DIR="$BUILD_DIR/$OS-$ARCH"
 
         if [[ "$OS" == "osx" ]]; then
-            APP_DIR="$ROOT_DIR/bin/$CONFIG/net6.0/$OS-$ARCH/rgit.app"
+            if [[ "$ARCH" == "arm64" ]]; then
+                cp "$ROOT_DIR"/Assets/lib/arm64/libgit2*.dylib "$BUILD_DIR/publish" || exit 1
+            fi
+            
+            APP_DIR="$BUILD_DIR/rgit.app"
             rm -rf "$APP_DIR"
             mkdir -p "$APP_DIR" || exit 1
-            cp -r "$ROOT_DIR"/Assets/osx/* "$APP_DIR" || exit 1
-            cp "$PUBLISH_DIR"/publish/* "$APP_DIR/Contents/MacOS" || exit 1
-            rm "$APP_DIR/Contents/MacOS/README"
-            cd "$PUBLISH_DIR" || exit 1
+            cp -r "$ROOT_DIR/Assets/osx/"/* "$APP_DIR" || exit 1
+            # Need to use stream syntax because sed on osx errors file input.
+            < "$ROOT_DIR/Assets/osx/Contents/Info.plist" sed -e "s/\${VERSION}/$VERSION/g" -e "s/\${VERSION_SHORT}/$VERSION_SHORT/g" > "$APP_DIR/Contents/Info.plist"
+            cp "$BUILD_DIR"/publish/* "$APP_DIR/Contents/MacOS" || exit 1
+            rm "$APP_DIR/Contents/MacOS/README" || exit 1
+            cd "$BUILD_DIR" || exit 1
             tar -czf "rgit.app.tgz" "rgit.app" || exit 1
         fi
     else
         dotnet build -c "$CONFIG" "${VERBOSITY[@]}" "$SHARED_COMPILATION"
+
+        if [[ "$OS" == "osx" ]] && [[ "$ARCH" == "arm64" ]]; then
+            cp "$ROOT_DIR"/Assets/osx/Contents/MacOS/arm64/libgit2-*.dylib "$BUILD_DIR/" || exit 1
+        fi
     fi
 }
